+++ include/H2DE/engine/H2DE_engine.h
@@ -47,6 +47,32 @@ class H2DE_ObjectManager;
 
 using H2DE_Delay = H2DE_Timeline;
 
+namespace H2DE {
+    /**
+     * @brief Creates and initializes the H2DE engine singleton instance using the provided configuration data.
+     * 
+     * This function can only be called once during the application lifetime. It initializes all core components of the engine,
+     * including the window, renderer, asset manager, camera, object manager, audio, timeline manager, and settings.
+     * 
+     * @param data Constant reference to H2DE_EngineData structure containing engine initialization parameters like window and camera settings.
+     * 
+     * @return Pointer to the created H2DE_Engine instance. Returns nullptr if creation fails or if the engine already exists.
+     * 
+     * @throws Throws a runtime error if an attempt is made to create more than one engine instance.
+     *         Exceptions are caught internally and reported via a MessageBox.
+     */
+    H2DE_Engine* createEngine(const H2DE_EngineData& data);
+    /**
+     * @brief Destroys the H2DE engine instance and releases all allocated resources.
+     * 
+     * Deletes all core components (window, renderer, asset loader, camera, timeline manager, etc.) 
+     * and clears all objects managed by the engine.
+     * 
+     * @param engine Pointer to the H2DE_Engine instance to destroy.
+     */
+    void destroyEngine(H2DE_Engine* engine);
+};
+
 /**
  * @class H2DE_Engine
  * @brief Central class managing the whole 2D game engine lifecycle and subsystems.
@@ -67,8 +93,8 @@ using H2DE_Delay = H2DE_Timeline;
  */
 class H2DE_Engine {
 public:
-    friend H2DE_Engine* H2DE_CreateEngine(const H2DE_EngineData& data);
-    friend void H2DE_DestroyEngine(H2DE_Engine* engine);
+    friend H2DE_Engine* H2DE::createEngine(const H2DE_EngineData& data);
+    friend void H2DE::destroyEngine(H2DE_Engine* engine);
 
     /**
      * @brief Starts the main loop of the engine, processing events, updating, and rendering frames.
@@ -86,15 +112,30 @@ public:
         isRunning = false;
     }
 
-    /**
-     * @brief Loads all assets (textures, sounds) from the specified directory into the engine.
+     /**
+     * @brief Loads all supported assets (textures and sounds) from a given directory.
      * 
-     * This function verifies that the directory exists and is valid, then imports files.
-     * It updates the engine's texture and sound caches, warning if any assets are overridden.
+     * This function scans the provided directory (recursively) and loads all valid image and audio files
+     * into the engine. It also updates the engine's internal caches for textures and sounds.
      * 
-     * @param directory Path to the folder containing assets to load.
+     * If a file with the same name already exists in the cache, it will be overridden (a warning is logged).
+     * 
+     * @param directory Path to the folder containing the assets to load.
      */
-    void loadAssets(const std::string& directory);
+    void loadAssetsSync(const std::string& directory);
+    /**
+     * @brief Loads all supported assets asynchronously from a given directory.
+     * 
+     * This function works the same as `loadAssetsSync` but performs the operation on a separate thread.
+     * It allows you to provide progress feedback (0.0f to 1.0f) and a callback when loading is complete.
+     * 
+     * The engine remains responsive during the loading process.
+     * 
+     * @param directory Path to the folder containing the assets to load.
+     * @param progress Callback function that receives the current progress as a float from 0.0f to 1.0f.
+     * @param completed Callback function called once all assets are loaded and stored in buffers.
+     */
+    void loadAssetsAsync(const std::string& directory, const std::function<void(float)>& progress, const std::function<void()>& completed);
     /**
      * @brief Loads or overrides a font with the given name in the engine's font collection.
      * 
@@ -214,38 +255,7 @@ public:
      * @warning Compilation will fail if `H2DE_Object_T` declares a `H2DE_DataType`, but no or more than one extra argument is provided.
      */
     template<typename H2DE_Object_T, typename... H2DE_SpecificObjectData_T>
-    H2DE_Object_T* createObject(const H2DE_ObjectData& objectData, H2DE_SpecificObjectData_T&&... specificObjectData) {
-        H2DE_Object_T* object = nullptr;
-
-        constexpr int nbArgs = sizeof...(H2DE_SpecificObjectData_T);
-        constexpr bool hasDataType = has_H2DE_DataType<H2DE_Object_T>::value;
-
-        if (nbArgs > 1) {
-            static_assert((hasDataType && nbArgs == 1) || (!hasDataType && nbArgs == 0), "Invalid arguments passed to createObject");
-        }
-        
-        constexpr bool objectHasNoSpecificDataType = (!hasDataType && nbArgs == 0);
-        constexpr bool objectHasSpecificDataType = (hasDataType && nbArgs == 1);
-
-        if constexpr (objectHasNoSpecificDataType) {
-            object = new H2DE_Object_T(this, objectData);
-
-        } else if constexpr (objectHasSpecificDataType) {
-            object = new H2DE_Object_T(this, objectData, std::forward<H2DE_SpecificObjectData_T>(specificObjectData)...);
-        
-        } else {
-            return nullptr;
-        }
-
-        objects.push_back(object);
-
-        constexpr bool isButtonObject = (std::is_same_v<H2DE_Object_T, H2DE_ButtonObject>);
-        if constexpr (isButtonObject) {
-            refreshObjectManager();
-        }
-
-        return object;
-    }
+    H2DE_Object_T* createObject(const H2DE_ObjectData& objectData, H2DE_SpecificObjectData_T&&... specificObjectData);
 
     /**
      * @brief Destroys and removes the specified object from the engine.
@@ -458,29 +468,5 @@ private:
 
     static bool isPositionGreater(H2DE_Object* a, H2DE_Object* b);
 };
-
-/**
- * @brief Creates and initializes the H2DE engine singleton instance using the provided configuration data.
- * 
- * This function can only be called once during the application lifetime. It initializes all core components of the engine,
- * including the window, renderer, asset manager, camera, object manager, audio, timeline manager, and settings.
- * 
- * @param data Constant reference to H2DE_EngineData structure containing engine initialization parameters like window and camera settings.
- * 
- * @return Pointer to the created H2DE_Engine instance. Returns nullptr if creation fails or if the engine already exists.
- * 
- * @throws Throws a runtime error if an attempt is made to create more than one engine instance.
- *         Exceptions are caught internally and reported via a MessageBox.
- */
-H2DE_Engine* H2DE_CreateEngine(const H2DE_EngineData& data);
-/**
- * @brief Destroys the H2DE engine instance and releases all allocated resources.
- * 
- * Deletes all core components (window, renderer, asset loader, camera, timeline manager, etc.) 
- * and clears all objects managed by the engine.
- * 
- * @param engine Pointer to the H2DE_Engine instance to destroy.
- */
-void H2DE_DestroyEngine(H2DE_Engine* engine);
 
 #endif