+++ src/utils/H2DE_vector2d.cpp
@@ -26,7 +26,7 @@ H2DE_Vector2D<H2DE_Vector2D_T>& H2DE_Vector2D<H2DE_Vector2D_T>::operator*=(float
 }
 
 template<typename H2DE_Vector2D_T>
-H2DE_Vector2D<H2DE_Vector2D_T>& H2DE_Vector2D<H2DE_Vector2D_T>::operator/=(float divider) noexcept {
+H2DE_Vector2D<H2DE_Vector2D_T>& H2DE_Vector2D<H2DE_Vector2D_T>::operator/=(float divider) {
     x /= divider;
     y /= divider;
     return *this;